PCI: qcom: Implement RC shutdown/power up
Currently, most (if not all) Qualcomm SoCs with devices connected over
PCIe are incapable of entering S3-equivalent sleep states, wasting tons
of power. That is in part because the PCIe host driver always keeps the
related resources online (see [1]) and partly because L1ss support is
not yet implemented (and that's something to solve separately).

Implement full link & controller shutdown and re-initialization
routines to let us disable all on-SoC data and power lines going to the
PCIe RC.

This is mainly intended for SC8280XP, which uniquely cannot keep the
link up, unless VDD_CX (the main system power rail, which is normally
shut down during platform suspend) is online, but can be reused on
other RPMh-based Qualcomm SoCs as a middleground before L1ss support
is introduced (or to achieve lower power numbers than with L1ss).

[1] Commit ad9b9b6e36c9 ("PCI: qcom: Add support for system suspend and resume")
Co-developed-by: Bjorn Andersson <quic_bjorande@quicinc.com>
Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>