PCI: dwc: Allow forcibly disabling the link on suspend
Certain platforms require that all PCIe hosts (and thereby links and
devices) are shut down completely, in order to achieve S3 (or similar)
sleep levels. This is adjacent to the issues seen on recent AMD and
Intel platforms (described in [1] and [2] respectively).

In such cases, runtime PM for the PCIe host goes out of the window
anyway, so resume latency is not exactly a thing to worry about.

Allow glue drivers to skip the L1ss check in dw_pcie_suspend_noirq()
and always fall into the link shutdown path.

[1] Commit 6485fc18faa0 ("ACPI: Add quirks for AMD Renoir/Lucienne CPUs to force the D3 hint")
[2] Commit df4f9bc4fb9c ("nvme-pci: add support for ACPI StorageD3Enable property")

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>