Allow nonposted-mmio on single devices
For a plethora of reasons, usually not very pretty, a single device on
a given bus/platform may require a different type of memory mapping.

With nonposted-mmio specifically, we want to limit the scope of setting
it where it's not necessary, so as not to shoot ourselves in the foot
in the performance department.

This series allows mapping a single device as nE on arm64

Signed-off-by: Konrad Dybcio <konrad.dybcio@oss.qualcomm.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20250315-topic-nonposted_mmio-5393a3de0d35",
    "prefixes": []
  }
}