OPP: of: Allow missing opp-supported-hw as an opt-in
Certain drivers are used with hardware that may or may not (varying by
platform) support a form of speed binning.

Currently, two things need to be in sync:

A: opp-supported-hw is present under an OPP node
B: the driver has called devm_pm_opp_set_supported_hw()

Now, it happens to be that A&B is fine and !A&!B is fine, but if only
one condition is met, _opp_is_supported() always returns false.

One could think that it should be as simple as only calling
devm_pm_opp_set_supported_hw() if there's a table of speed bins present
in the driver, however we also have to consider DT backwards (and to
some extent forwards) compatibility, meaning A can be both true and
false, depending on the FDT supplied.

Moreover, certain IP blocks that are carbon-copied across a number of
SoCs may only feature multiple flavors on some, but not others -
meaning opp-supported-hw presence could vary even in the "new enough
DT" case.

Allow drivers to fall back to behavior removed by Commit a4ee4545932d
("PM / OPP: Don't support OPP if it provides supported-hw but platform
does not"), where "old" DTs (lacking opp-supported-hw) continue to be
supported by the driver that has been since taught about multiple
possible speed bins / variants of an IP block existing.

Signed-off-by: Konrad Dybcio <konrad.dybcio@oss.qualcomm.com>