drm/msm/adreno: Allow opp-supported-hw to be absent
Some platforms only offer a single frequency plan (i.e. no discrete
speed bins). To avoid having to come up with meaningless blurb to fill
in the aforementioned property, and to allow for supplying a bin table
to the driver after the initial submission, assume that if
opp-supported-hw is missing, all OPPs shall be allowed.

Signed-off-by: Konrad Dybcio <konrad.dybcio@oss.qualcomm.com>